+++ packages/showcases/svelte/stories/components/overlays/VtmnPopover/VtmnPopover.stories.svelte
@@ -18,7 +18,7 @@
   <VtmnPopover position={args.position} id={args.id}>
     <svelte:fragment slot="title">{args.title}</svelte:fragment>
     <svelte:fragment slot="body">{args.body}</svelte:fragment>
-    <VtmnLink slot="placeholder">Popover on the {args.position}</VtmnLink>
+    <VtmnLink>Popover on the {args.position}</VtmnLink>
   </VtmnPopover>
 </Template>