+++ packages/sources/vue/src/components/index.ts
@@ -6,10 +6,13 @@ import VtmnBadge from './indicators/VtmnBadge';
 import VtmnLoader from './indicators/VtmnLoader';
 import VtmnPrice from './indicators/VtmnPrice';
 import VtmnTag from './indicators/VtmnTag';
+import VtmnAlert from './overlays/VtmnAlert';
+import VtmnTooltip from './overlays/VtmnTooltip';
 import VtmnCheckbox from './selection-controls/VtmnCheckbox';
 import VtmnChip from './selection-controls/VtmnChip';
 import VtmnRadioButton from './selection-controls/VtmnRadioButton';
 import VtmnToggle from './selection-controls/VtmnToggle';
+import VtmnAccordion from './structure/VtmnAccordion';
 import VtmnDivider from './structure/VtmnDivider';
 import VtmnSkeleton from './structure/VtmnSkeleton';
 
@@ -25,7 +28,10 @@ export {
   VtmnChip,
   VtmnRadioButton,
   VtmnTag,
+  VtmnAlert,
+  VtmnTooltip,
   VtmnToggle,
+  VtmnAccordion,
   VtmnDivider,
   VtmnSkeleton,
 };