+++ packages/sources/vue/src/index.ts
@@ -36,6 +36,7 @@ export { default as VtmnTag } from './components/indicators/VtmnTag';
 // Navigation
 
 // Overlays
+export { default as VtmnAlert } from './components/overlays/VtmnAlert';
 export { default as VtmnTooltip } from './components/overlays/VtmnTooltip';
 
 // Selection controls
@@ -45,5 +46,6 @@ export { default as VtmnRadioButton } from './components/selection-controls/Vtmn
 export { default as VtmnToggle } from './components/selection-controls/VtmnToggle';
 
 // Structure
+export { default as VtmnAccordion } from './components/structure/VtmnAccordion';
 export { default as VtmnDivider } from './components/structure/VtmnDivider';
 export { default as VtmnSkeleton } from './components/structure/VtmnSkeleton';